feat: moving authentication logic to the authenticationService
